Fix type conversion warning

+++ OPHD/ResourcePool.cpp
@@ -27,7 +27,7 @@ ResourcePool::ResourcePool(int cmo, int cmno, int rmo, int rmno, int cm, int cmn
 }
 
 
-ResourcePool::ResourcePool(size_t capacity) : _capacity(capacity)
+ResourcePool::ResourcePool(size_t capacity) : _capacity(static_cast<int>(capacity))
 {}
